Modify tests namespaces
