+++ GovLib.Tests/ProPublica/Congress/Bills/Fixtures/BillsByMemberFixture.cs
@@ -2,7 +2,7 @@ using System.Threading;
 using GovLib.Contracts;
 using GovLib.ProPublica;
 
-namespace GovLib.Tests.ProPublicaTests.CongressTests.BillsTests
+namespace GovLib.Tests.ProPublica.Congress.Bills
 {
     public class BillsByMemberFixture : CongressFixture
     {
