+++ GovLib.Tests/ProPublica/Congress/CongressFixture.cs
@@ -2,17 +2,17 @@ using System;
 using System.Threading;
 using GovLib.ProPublica;
 
-namespace GovLib.Tests.ProPublicaTests.CongressTests
+namespace GovLib.Tests.ProPublica.Congress
 {
     public abstract class CongressFixture
     {
         public string ApiKey { get; }
-        public Congress Congress { get; }
+        public GovLib.ProPublica.Congress Congress { get; }
 
         public CongressFixture()
         {
             ApiKey = Environment.GetEnvironmentVariable("PROPUBLICA_API_KEY");
-            Congress = new Congress(ApiKey);
+            Congress = new GovLib.ProPublica.Congress(ApiKey);
         }
     }
 }
