+++ GovLib.Tests/ProPublica/Congress/Members/Fixtures/AllSenatorsFixture.cs
@@ -2,7 +2,7 @@ using System.Threading;
 using GovLib.Contracts;
 using GovLib.ProPublica;
 
-namespace GovLib.Tests.ProPublicaTests.CongressTests.MembersTests
+namespace GovLib.Tests.ProPublica.Congress.Members
 {
     public class AllSenatorsFixture : CongressFixture
     {
