+++ GovLib.Tests/ProPublica/Congress/Members/GetAllRepresentativesTests.cs
@@ -1,6 +1,6 @@
 using Xunit;
 
-namespace GovLib.Tests.ProPublicaTests.CongressTests.MembersTests
+namespace GovLib.Tests.ProPublica.Congress.Members
 {
     public class GetAllRepresentativesTests : IClassFixture<AllReprentativesFixture>
     {
