+++ GovLib.Tests/ProPublica/Congress/Members/GetRepresentativeFromDistrictTests.cs
@@ -1,6 +1,6 @@
 using Xunit;
 
-namespace GovLib.Tests.ProPublicaTests.CongressTests.MembersTests
+namespace GovLib.Tests.ProPublica.Congress.Members
 {
     public class GetRepresentativesFromDistrictTests : IClassFixture<RepresentativeFromDistrictFixture>
     {
