+++ GovLib.Tests/ProPublica/Congress/Members/GetRepresentativesByStateTests.cs
@@ -1,13 +1,13 @@
 using Xunit;
 
-namespace GovLib.Tests.ProPublicaTests.CongressTests.MembersTests
+namespace GovLib.Tests.ProPublica.Congress.Members
 {
-    public class GetSenatorsByStateTests : IClassFixture<SenatorsByStateFixture>
+    public class GetRepresentativesByStateTests : IClassFixture<RepresentativesByStateFixture>
     {
         
-        public SenatorsByStateFixture Fixture { get; }
+        public RepresentativesByStateFixture Fixture { get; }
 
-        public GetSenatorsByStateTests(SenatorsByStateFixture fixture)
+        public GetRepresentativesByStateTests(RepresentativesByStateFixture fixture)
         {
             this.Fixture = fixture;
         }
@@ -15,61 +15,61 @@ namespace GovLib.Tests.ProPublicaTests.CongressTests.MembersTests
         [Fact]
         public void CollectionIsNotNull()
         {
-            Assert.NotNull(Fixture.StateSenators);
+            Assert.NotNull(Fixture.StateReps);
         }
 
         [Fact]
         public void CollectionIsNotEmpty()
         {
-            Assert.NotEmpty(Fixture.StateSenators);
+            Assert.NotEmpty(Fixture.StateReps);
         }
 
         [Fact]
         public void MemberCardsAreNotNull()
         {
-            foreach (var member in Fixture.StateSenators)
+            foreach (var member in Fixture.StateReps)
                 Assert.NotNull(member);
         }
 
         [Fact]
         public void MemberCardsHaveFirstName()
         {
-            foreach (var member in Fixture.StateSenators)
+            foreach (var member in Fixture.StateReps)
                 Assert.False(string.IsNullOrEmpty(member.FirstName));
         }
 
         [Fact]
         public void MemberCardsHaveLastName()
         {
-            foreach (var member in Fixture.StateSenators)
+            foreach (var member in Fixture.StateReps)
                 Assert.False(string.IsNullOrEmpty(member.LastName));
         }
 
         [Fact]
         public void MemberCardsHaveFullName()
         {
-            foreach (var member in Fixture.StateSenators)
+            foreach (var member in Fixture.StateReps)
                 Assert.False(string.IsNullOrEmpty(member.FullName));
         }
 
         [Fact]
         public void MemberCardsHaveAnID()
         {
-            foreach (var member in Fixture.StateSenators)
+            foreach (var member in Fixture.StateReps)
                 Assert.False(string.IsNullOrEmpty(member.CongressID));
         }
 
         [Fact]
         public void MemberCardsHaveAHomeState()
         {
-            foreach (var member in Fixture.StateSenators)
+            foreach (var member in Fixture.StateReps)
                 Assert.NotNull(member.State);
         }
 
         [Fact]
         public void MemberCardsHaveAParty()
         {
-            foreach (var member in Fixture.StateSenators)
+            foreach (var member in Fixture.StateReps)
                 Assert.False(string.IsNullOrEmpty(member.Party));
         }
     }
