+++ GovLib.Tests/ProPublica/Congress/Members/GetSenatorsLeavingTests.cs
@@ -1,6 +1,6 @@
 using Xunit;
 
-namespace GovLib.Tests.ProPublicaTests.CongressTests.MembersTests
+namespace GovLib.Tests.ProPublica.Congress.Members
 {
     public class GetSenatorsLeavingTests : IClassFixture<SenatorsLeavingFixture>
     {
